Move ProcessedPeak and SupportingPeak classes out of AnalysisResult.
